+++ engine/api/src/main/java/pl/touk/nussknacker/engine/api/DefaultValue.java
@@ -1,0 +1,12 @@
+package pl.touk.nussknacker.engine.api;
+
+import java.lang.annotation.ElementType;
+import java.lang.annotation.Retention;
+import java.lang.annotation.RetentionPolicy;
+import java.lang.annotation.Target;
+
+@Target(ElementType.PARAMETER)
+@Retention(RetentionPolicy.RUNTIME)
+public @interface DefaultValue {
+    String value();
+}