+++ engine/interpreter/src/main/scala/pl/touk/nussknacker/engine/definition/parameter/OptionalDeterminer.scala
@@ -1,0 +1,10 @@
+package pl.touk.nussknacker.engine.definition.parameter
+
+import javax.annotation.Nullable
+
+object OptionalDeterminer {
+
+  def isOptional(parameterData: ParameterData, isScalaOptionParameter: Boolean, isJavaOptionalParameter: Boolean): Boolean =
+    isScalaOptionParameter || isJavaOptionalParameter || parameterData.getAnnotation[Nullable].isDefined
+
+}