+++ engine/interpreter/src/main/scala/pl/touk/nussknacker/engine/definition/parameter/defaults/ConfigParameterDefaultValueDeterminer.scala
@@ -1,0 +1,9 @@
+package pl.touk.nussknacker.engine.definition.parameter.defaults
+
+object ConfigParameterDefaultValueDeterminer extends ParameterDefaultValueDeterminer {
+
+  override def determineParameterDefaultValue(parameters: DefaultValueDeterminerParameters): Option[String] = {
+    parameters.parameterConfig.defaultValue
+  }
+
+}