+++ engine/interpreter/src/main/scala/pl/touk/nussknacker/engine/definition/parameter/validator/EditorBasedValidatorExtractor.scala
@@ -1,15 +1,15 @@
 package pl.touk.nussknacker.engine.definition.parameter.validator
 
 import pl.touk.nussknacker.engine.api.definition._
-import pl.touk.nussknacker.engine.definition.parameter.editor.EditorExtractor
 
 object EditorBasedValidatorExtractor extends ValidatorExtractor {
 
   override def extract(params: ValidatorExtractorParameters): Option[ParameterValidator] = {
-    EditorExtractor.extract(params.parameterData, params.parameterConfig) match {
-      case Some(FixedValuesParameterEditor(possibleValues)) => Some(FixedValuesValidator(possibleValues))
-      case Some(JsonParameterEditor) => Some(JsonValidator)
+    params.extractedEditor.flatMap {
+      case FixedValuesParameterEditor(possibleValues) => Some(FixedValuesValidator(possibleValues))
+      case JsonParameterEditor => Some(JsonValidator)
       case _ => None
     }
   }
+
 }