+++ engine/interpreter/src/main/scala/pl/touk/nussknacker/engine/definition/parameter/validator/ValidatorExtractor.scala
@@ -2,7 +2,6 @@ package pl.touk.nussknacker.engine.definition.parameter.validator
 
 import pl.touk.nussknacker.engine.api.definition.{ParameterEditor, ParameterValidator}
 import pl.touk.nussknacker.engine.api.process.ParameterConfig
-import pl.touk.nussknacker.engine.api.typed.typing.TypingResult
 import pl.touk.nussknacker.engine.definition.parameter.ParameterData
 
 trait ValidatorExtractor {
@@ -13,4 +12,5 @@ trait ValidatorExtractor {
 
 case class ValidatorExtractorParameters(parameterData: ParameterData,
                                         isOptional: Boolean,
-                                        parameterConfig: ParameterConfig)
+                                        parameterConfig: ParameterConfig,
+                                        extractedEditor: Option[ParameterEditor])